Set up config file to include basic information and pass it to Modal
We set up information such as colors, text, and structure in the config
file so that information can be passed down to various tools. We also
successfully passed some information to the FHModal.

Basic initial structure of config object for modal is complete

Object now passes in correct parameters to FHModal

Setup established