Create container element to better position elements in SearchInput
This commit creates a container element to better position the various
elements in SearchInput. For instance, it is now much easier to center
the items on the page. The container element, like the others, can be
modified in FH Config and also includes a debug state which shows a red
border.

Create container that holds all elements for SearchInput to position them

fix minor issue with hover state on details button

disables full details button if no event is selected

Remove console log

Consolidate global vars within SearchInput