Gn search input expansion (#20)
Remove event listeners to clear input fields on click

Change 'Where To? text for event picker to 'Which Event?'

* Change old input field names to LeftInputField and RightInputField

Add framework for details input field and fix bug with right input field

* Add option to include or exclude details input field

* Abandoning details dropdown

* Eliminate refs to details field in config

* Renames Go Button and changes naming for new details button

Details button now toggles whether or not it exists

Format display button

Full event button works

* create hove state styling for details button; need to refactor this later

Create container element to better position elements in SearchInput

* This commit creates a container element to better position the various
elements in SearchInput. For instance, it is now much easier to center
the items on the page. The container element, like the others, can be
modified in FH Config and also includes a debug state which shows a red
border.

* Create container that holds all elements for SearchInput to position them

* fix minor issue with hover state on details button

* disables full details button if no event is selected

* Remove console log

* Consolidate global vars within SearchInput