Wrap Go action button in an anchor tag to permit lightframes
This commit wraps the main action button in an anchor tag in order to
allow the lightframes to activate. This required restructuring some
event listeners and making API calls each time the state of one of the
input fields changes. When this happens, it reassigns the value of the
href attribute for the button wrapper link so that the button knows to
take the user to the correct event and date.

WIP -- adds a URL var that keeps track of the URL to go to when user clicks the Go button

Button functionality restored for event without date

Add functionality for if there is a date but no event specified

WIP -- working on logic for date picker

Fix bug with datePicker

WIP -- still fighting bug

finally got all functionality restored for the a tag lightframe refactor

Fix bug with date picker

The date picker was sometimes not updating the url for the button link
properly. I fixed this by making the targetID, which records the ID of
the currently selected event, be null when no event is selected.