Integrator can impose horizontal or vertical orientation for input
One can now pass a horizontal or vertical specification to the config
object, which dictates the orientation of the input field elements. For
instance, the vertical spec causes the items to appear as a stack of
block elements as in a mobile format, while the horizontal spec causes
the items to appear in a more desktop-friendly mode.

Add display options for right input field

Solve regex problem with March v. May

WIP -- formats left and right input field margins based on stack orientation

Format details button for vertical and horizontal stack cases

Finished with horiz and vertical stacking for input; now need to work on responsiveness