Gn responsiveness and lightframe fix (#30)
Add comments and notes from talk with Adam

Reformats footer element positioning according to number of elements

In contrast to the prior situation of each footer element having its
own margin and padding options, this restructuring automatically formats
each element in the footer according to how many elements there are. For
instance, if there is only one element, it is centered in the footer and
its text is enlarged in order to make it appear more prominently. This
commit also adds a hover state to each element and adds a dynamic
calendar link to the second element.

Simplify object specs for footer

Remove margins from styling for each item in footer

Add unique class names to each element in footer

Change class to id for each element in footer

Format footer with single element

Format footer with two elements

Format footer with three elements

Format footer with four elements

Add dynamic calendar link to footer second element

Add dynamic month and year to url link for calendar second element

Add hover state to each element

Finished with footer restructuring for now

* Wrap Go action button in an anchor tag to permit lightframes

This commit wraps the main action button in an anchor tag in order to
allow the lightframes to activate. This required restructuring some
event listeners and making API calls each time the state of one of the
input fields changes. When this happens, it reassigns the value of the
href attribute for the button wrapper link so that the button knows to
take the user to the correct event and date.

WIP -- adds a URL var that keeps track of the URL to go to when user clicks the Go button

Button functionality restored for event without date

Add functionality for if there is a date but no event specified

WIP -- working on logic for date picker

Fix bug with datePicker

WIP -- still fighting bug

finally got all functionality restored for the a tag lightframe refactor

Fix bug with date picker

The date picker was sometimes not updating the url for the button link
properly. I fixed this by making the targetID, which records the ID of
the currently selected event, be null when no event is selected.

* Wrap details button in anchor tag to permit lightframes

Like with the prior commit, this commit wraps a button in an anchor tag
to allow the lightframe to generate.

Restore functionality to details button with a tag

* Set more default styles for search input elements

A number of attributes, such as margins and padding, are now set by
default and no longer customizable via the config file for the search
input elements.

Make the Which Event and When referents dynamic in the event listeners

Finishes restyling margins and padding

* Integrator can impose horizontal or vertical orientation for input

One can now pass a horizontal or vertical specification to the config
object, which dictates the orientation of the input field elements. For
instance, the vertical spec causes the items to appear as a stack of
block elements as in a mobile format, while the horizontal spec causes
the items to appear in a more desktop-friendly mode.

Add display options for right input field

Solve regex problem with March v. May

WIP -- formats left and right input field margins based on stack orientation

Format details button for vertical and horizontal stack cases

Finished with horiz and vertical stacking for input; now need to work on responsiveness

* Add responsiveness to search input

The search input now automatically renders in a vertically-stacked,
mobile-friendly way when the viewport is below 700px in width.

Make the search input responsive

* Make footer responsive

The footer displays vertically rather than horizontally of the viewport
width is below 700px.

Finish footer responsiveness