Clean up configuration of tools by removing unneeded items
This commit removes several unneeded items, such as button height and
width, from the FHConfig file. It also modifies the search input object
in FHConfig slightly to move the "stack orientation" property to the
root of the object, since it affects both the input fields and the
buttons.

Move stack orientation object to root of Search Input object

Remove buttons height and width as customizable properties

Remove unneeded item from footer in FHToolCreator

Remove height and width as customizable elements for footer and footer elements