Refactor to make tests more robust (#73)
Tests are now more robust and flexible. For instance, rather than
depending on the right input field to be the date dropdown and the left
input field to be the event dropdown, there are now classes that are
used for testing those fields. Also, I created a way to skip tests if a
condition is not met (for instance, if a text node in the footer does
not exist.)

Refactors search input tests; all passing

Add eventClass and dateClass to the input fields

Add name and event id categories for each dropdown to make testing less brittle

Add conditional to relevant search input tests if shortname is bodyglove

New means of conditionally ignoring tests if no bodyglove works

Clean up code in tests

Throws new error when both input fields are datePicker or eventPicker

Add fix for footer test is third text node does not exist

Footer is done

Final code touchups; all search input tests passing